Add storage and namespace cost helpers
