+++ ledger/block/src/helpers/mod.rs
@@ -12,5 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod cost;
+pub use cost::*;
+
 mod target;
 pub use target::*;
